Curate application website to include schema
Application websites seems to be input with or without schema by
users. Civstack needs was asuming schema (http://) was always
present and links are created in the application based on that
asumption.

This might look like a temporary patch. When the received list
of approved applications is translated a function call was
added to ensure all urls look alike.